requirements.all.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --extra=all --output-file=requirements.all.txt
#
altgraph==0.17.2
    # via pyinstaller
attrs==21.4.0
    # via environ-config
black==22.1.0
    # via qBitrr (setup.py)
bleach==4.1.0
    # via readme-renderer
bump2version==1.0.1
    # via qBitrr (setup.py)
cachetools==4.2.4
    # via qBitrr (setup.py)
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   black
    #   pip-tools
colorama==0.4.4
    # via
    #   qBitrr (setup.py)
    #   twine
coloredlogs==15.0.1
    # via qBitrr (setup.py)
cryptography==36.0.1
    # via secretstorage
dill==0.3.4
    # via
    #   multiprocess
    #   pathos
distlib==0.3.4
    # via virtualenv
docutils==0.18.1
    # via readme-renderer
environ-config==21.2.0
    # via qBitrr (setup.py)
ffmpeg-python==0.2.0
    # via qBitrr (setup.py)
filelock==3.4.2
    # via virtualenv
future==0.18.2
    # via ffmpeg-python
humanfriendly==10.0
    # via coloredlogs
identify==2.4.7
    # via pre-commit
idna==3.3
    # via requests
importlib-metadata==4.10.1
    # via
    #   keyring
    #   twine
isort==5.10.1
    # via qBitrr (setup.py)
jaraco.context==4.1.1
    # via jaraco.docker
jaraco.docker==2.0
    # via qBitrr (setup.py)
jaraco.functools==3.5.0
    # via jaraco.docker
jeepney==0.7.1
    # via
    #   keyring
    #   secretstorage
keyring==23.5.0
    # via twine
more-itertools==8.12.0
    # via jaraco.functools
multiprocess==0.70.12.2
    # via pathos
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
packaging==21.3
    # via
    #   bleach
    #   qBitrr (setup.py)
pathos==0.2.8
    # via qBitrr (setup.py)
pathspec==0.9.0
    # via black
peewee==3.14.7
    # via qBitrr (setup.py)
pep517==0.12.0
    # via pip-tools
ping3==3.0.2
    # via qBitrr (setup.py)
pip-tools==6.4.0
    # via qBitrr (setup.py)
pkginfo==1.8.2
    # via twine
platformdirs==2.4.1
    # via
    #   black
    #   virtualenv
pox==0.3.0
    # via pathos
ppft==1.6.6.4
    # via pathos
pre-commit==2.17.0
    # via qBitrr (setup.py)
pyarr==2.0.6
    # via qBitrr (setup.py)
pycparser==2.21
    # via cffi
pygments==2.11.2
    # via readme-renderer
pyinstaller==4.8
    # via qBitrr (setup.py)
pyinstaller-hooks-contrib==2022.0
    # via pyinstaller
pyparsing==3.0.7
    # via packaging
pyupgrade==2.31.0
    # via qBitrr (setup.py)
pyyaml==6.0
    # via pre-commit
qbittorrent-api==2021.8.23
    # via qBitrr (setup.py)
readme-renderer==32.0
    # via twine
requests==2.26.0
    # via
    #   pyarr
    #   qBitrr (setup.py)
    #   qbittorrent-api
    #   requests-toolbelt
    #   twine
requests-toolbelt==0.9.1
    # via twine
rfc3986==2.0.0
    # via twine
secretstorage==3.3.1
    # via keyring
six==1.16.0
    # via
    #   bleach
    #   ppft
    #   qbittorrent-api
    #   virtualenv
tokenize-rt==4.2.1
    # via pyupgrade
toml==0.10.2
    # via pre-commit
tomli==2.0.0
    # via
    #   black
    #   pep517
tomlkit==0.7.2
    # via qBitrr (setup.py)
tqdm==4.62.3
    # via twine
twine==3.7.1
    # via qBitrr (setup.py)
typing-extensions==4.0.1
    # via black
ujson==5.1.0
    # via qBitrr (setup.py)
urllib3==1.26.8
    # via
    #   qbittorrent-api
    #   requests
virtualenv==20.13.0
    # via pre-commit
webencodings==0.5.1
    # via bleach
wheel==0.37.1
    # via pip-tools
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
